requirements.txt
flask
libvirt-python